Mini Allies' Sheet
- Spawn interactive Scrolls
  - 6 Allies' CS
  - Pack of Scrolls: "Bestiary"
- Can Interact with mini sheets to get the focus of the correct entity
  - TODO: Browse among sheets (arrows)
- Fix: Total number of unit in CombatResources is now correct from the beginning
- Describe Bug: several units clicked

+++ src/constants.rs
@@ -57,7 +57,7 @@ pub mod combat {
     }
 
     pub mod skill {
-        pub const MAX_PARTY: i32 = 6;
+        pub const MAX_PARTY: usize = 6;
         pub const BAM: i32 = 150;
     }
